gpio: mvebu: use chained_irq_{enter,exit} for GIC compatibility
The GPIO block used on Marvell EBU SoCs is not only used connected to
the Marvell MPIC, but also connected to the ARM GIC, which uses the
fasteoi flow handler. In order to ensure that the eoi hook of the GIC
driver gets called, the GPIO driver should call chained_irq_enter()
and chained_irq_exit() in its handler. Without this, the first GPIO
interrupt locks up the system because it doesn't get acked at the GIC
level.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>