ARM: mvebu: enable the SDIO unit on Armada 375 DB
The Armada 375 DB board has a SD card slot connected to the SDIO unit
of the SoC, so this commit adds the relevant bits of Device Tree
informations to enable the usage of the SD card slot. In addition to
the SDIO unit itself, it also uses two GPIOs, one for card detect, and
one for write protect.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>