ata: add AHCI driver for the Marvell platform
This driver is similar to the ahci_platform driver, but adds the
Marvell-specific logic (especially Mbus window configuration) around
it.

For the moment, it duplicates the ahci_platform code, which isn't very
nice. Other approaches used are to make the ahci_<soc> driver register
a platform device that will trigger the ->probe() of the ahci_platform
driver, but this isn't nice easier. There is some discussion going on
to improve this situation.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>