ARM: mvebu: Enable NAND controller in Armada 370 Mirabox
The Armada 370 Mirabox has a NAND flash, so enable it in the devicetree
and add the partitions as prepared in the factory images.

In order to skip the driver's custom device detection and use only
ONFI detection, the marvell,keep-config parameter is used.
This is needed because we have no support for setting the timings parameters yet.

Change-Id: I3609b2a49b4015cf151ccc12826d6531396cad63
Signed-off-by: Ezequiel Garcia <ezequiel.garcia@free-electrons.com>
Signed-off-by: Jason Cooper <jason@lakedaemon.net>
Signed-off-by: Nadav Haklai <nadavh@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/5608
Tested-by: Star_Automation <star@marvell.com>